requirements.txt
# Core dependencies
pandas==2.0.0
numpy==1.24.3
matplotlib==3.7.1
gym==0.26.0
stable-baselines3==2.0.0
torch==2.0.1
tensorboard==2.13.0

# API and web framework
flask==2.3.2
flask-cors==3.0.10
flask-restful==0.3.10
flask-jwt-extended==4.5.2
flask-login==0.6.2
requests-oauthlib==1.3.1

# Data processing
yfinance==0.2.18
alpha_vantage==2.3.1
# ta-lib is installed separately via install_talib.py
scikit-learn==1.2.2

# Dashboard
plotly==5.14.1
dash==2.9.3
dash-uploader==0.6.0

# Testing
pytest==7.3.1

# Utilities
python-dotenv==1.0.0
requests==2.30.0
tqdm==4.65.0

# Cloud deployment
gunicorn==20.1.0

# AI Integration
google-generativeai==0.3.1
pillow==9.5.0
python-magic==0.4.27